Add files via upload
ll modern websites and web applications are built using three fundamental technologies: HTML, CSS, and JavaScript. In this project work, I focus on HTML and CSS.

HTML stands for HyperText Markup Language. It's a markup language that web developers use to structure and describe the content of a webpage (not a programming language).

HTML consists of elements that describe different types of content: paragraphs, links, headings, images, videos, etc. Web browsers understand HTML and render HTML code as websites.

In HTML, each element is made up of 3 parts:

The opening tag
The closing tag
The actual element
You can learn more at the MDN Web Docs.

CSS stands for Cascading Style Sheets language and is used to stylize elements written in a markup language such as HTML. It separates the content from the visual representation of the site. The relation between HTML and CSS is strongly tied together since HTML is the very foundation of a site and CSS is all of the aesthetics of an entire website.